requirements.txt
langchain
langchain-community
langchain-openai
openai
streamlit
requests
beautifulsoup4
lxml
python-dotenv